all guards and function done!

+++ src/controllers/stoneController.js
@@ -73,10 +73,10 @@ router.get("/:stoneId/details", async (req, res) => {
 router.get("/:stoneId/like", isAuth, async (req, res) => {
   const stoneId = req.params.stoneId;
   const userId = req.user._id;
+  const stone = await stoneService.getOne(stoneId);
 
-  const isOwner = isStoneOwner(stoneId, userId);
-  if (isOwner) {
-    res.redirect("/404");
+  if (stone.owner == stoneId) {
+    return res.redirect("/404");
   }
 
   try {
@@ -94,10 +94,10 @@ router.get("/:stoneId/like", isAuth, async (req, res) => {
 router.get("/:stoneId/delete", isAuth, async (req, res) => {
   const stoneId = req.params.stoneId;
   const userId = req.user._id;
+  const stone = await stoneService.getOne(stoneId);
 
-  const isOwner = isStoneOwner(stoneId, userId);
-  if (!isOwner) {
-    res.redirect("404");
+  if (!(stone.owner == userId)) {
+    return res.redirect("/404");
   }
 
   await stoneService.remove(stoneId);
@@ -111,10 +111,10 @@ router.get("/:stoneId/delete", isAuth, async (req, res) => {
 router.get("/:stoneId/edit", isAuth, async (req, res) => {
   const stoneId = req.params.stoneId;
   const userId = req.user._id;
+  const stone = await stoneService.getOne(stoneId);
 
-  const isOwner = isStoneOwner(stoneId, userId);
-  if (!isOwner) {
-    res.redirect("404");
+  if (!(stone.owner.toString() === userId)) {
+    return res.redirect("/404");
   }
 
   try {
@@ -130,10 +130,10 @@ router.post("/:stoneId/edit", isAuth, async (req, res) => {
   const stoneId = req.params.stoneId;
   const stoneData = req.body;
   const userId = req.user._id;
+  const stone = await stoneService.getOne(userId);
 
-  const isOwner = isStoneOwner(stoneId, userId);
-  if (!isOwner) {
-    res.redirect("404");
+  if (!(stone.owner == userId)) {
+    return res.redirect("/404");
   }
 
   try {
@@ -155,15 +155,5 @@ router.get("/search", async (req, res) => {
   const stones = await stoneService.getAll(filter).lean();
   res.render("stones/search", { stones, filter });
 });
-
-/*#############################
-########### HELPER #############
-################################*/
-async function isStoneOwner(stoneId, userId) {
-  const stone = await stoneService.getOne(stoneId);
-  const isOwner = userId === stone._id;
-
-  return isOwner;
-}
 
 export default router;
